pyproject.toml
[project]
name = "kang"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.29.0",
    "black>=25.1.0",
    "fastapi>=0.116.1",
    "flake8>=7.3.0",
    "flake8-pyproject>=1.2.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "python-rapidjson>=1.10",
    "redis>=6.2.0",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.42",
    "uvicorn[standard]>=0.35.0",
    # RAG 관련 의존성 (2024-2025 최신 버전)
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "sentence-transformers>=2.2.2",
    "huggingface-hub>=0.19.0",
    "tiktoken>=0.5.0",
    "pypdf>=3.17.0",
    "python-docx>=1.0.1",
    "pgvector>=0.2.0",
    "faiss-cpu>=1.7.4",
    "python-magic>=0.4.27",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
skip_gitignore = true

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
skip-string-normalization = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers"
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]

[tool.flake8]
max-line-length = 88
ignore = ["E501", "W503", "F722", "E203"]
per-file-ignores = [
    "tests/conftest.py:F401"
]
exclude = [
    "var",
    ".github",
    ".venv",
    "migrations"
]